REF: crear clase abstracta para encapsular lectura de certificados y poder reutilizar codigo en usuarios y compañias
